catch invalid_file

+++ app.rb
@@ -265,16 +265,6 @@
 		key :unique_values, Array
 		key :possible_names, Array
 		timestamps!
-		@@statuses = {
-			deleting: "deleting",
-			invalid_file: "invalid_file",
-			found_unique_values: "found_unique_values",
-			found_possible_matches: "found_possible_matches",
-			csv_is_ready: "csv_is_ready"
-		}
-		def self.statuses
-			@@statuses
-		end
 		after_create :set_field_names
 		def set_field_names
 			thread = Thread.new do
@@ -292,11 +282,14 @@
 			thread
 		end
 		def delete_in_5_minutes!
-			if self.status != "deleting"
-				self.update_attributes! status:  "deleting"
+			if self.status != "deleting" && self.status != "invalid_file"
+				
+				if self.status != "invalid_file"
+					self.update_attributes! status:  "deleting"
+				end
 				Thread.new do
 					sleep(5.mins)
-					if self && self.status == 'deleting'	
+					if self && (self.status == 'deleting' || self.status == "invalid_file")
 						self.destroy
 					end
 				end
